Fixed IniConfig to use streams, this makes it possible to test them. [skip ci]

+++ Dapplo.Config.Test/TestInterfaces/IIniTest.cs
@@ -36,7 +36,7 @@ namespace Dapplo.Config.Test.TestInterfaces {
 			set;
 		}
 		[DefaultValue(21), DataMember(EmitDefaultValue=true)]
-		int Age {
+		long Age {
 			get;
 			set;
 		}
